move createclaim  and verifyclaim to native
