+++ src/main/java/com/github/ontio/smartcontract/neovm/OntIdTx.java
@@ -798,55 +798,6 @@ public class OntIdTx {
         }
     }
 
-    /**
-     *
-     * @param signerOntid
-     * @param password
-     * @param context
-     * @param claimMap
-     * @param metaData
-     * @return
-     * @throws Exception
-     */
-    public String createOntIdClaim(String signerOntid, String password, String context, Map<String, Object> claimMap, Map metaData,Map clmRevMap,long expire) throws Exception {
-        Claim claim = null;
-        Map contentMap = sortMap(claimMap);
-
-        try {
-            String sendDid = (String) metaData.get("Issuer");
-            String receiverDid = (String) metaData.get("Subject");
-            if (sendDid == null || receiverDid == null) {
-                throw new SDKException(ErrorCode.DidNull);
-            }
-            String issuerDdo = sdk.nativevm().ontId().sendGetDDO(sendDid);
-            JSONArray owners = JSON.parseObject(issuerDdo).getJSONArray("Owners");
-            if (owners == null) {
-                throw new SDKException(ErrorCode.NotExistCliamIssuer);
-            }
-            String pubkeyId = null;
-            com.github.ontio.account.Account acct = sdk.getWalletMgr().getAccount(signerOntid, password);
-            String pk = Helper.toHexString(acct.serializePublicKey());
-            for (int i = 0; i < owners.size(); i++) {
-                JSONObject obj = owners.getJSONObject(i);
-                if (obj.getString("Value").equals(pk)) {
-                    pubkeyId = obj.getString("PubKeyId");
-                    break;
-                }
-            }
-            if (pubkeyId == null) {
-                throw new SDKException(ErrorCode.NotFoundPublicKeyId);
-            }
-            String[] receiverDidStr = receiverDid.split(":");
-            if (receiverDidStr.length != 3) {
-                throw new SDKException(ErrorCode.DidError);
-            }
-            metaData = sortMap(metaData);
-            claim = new Claim(sdk.getWalletMgr().getSignatureScheme(), acct, context, claimMap, metaData,clmRevMap,pubkeyId,expire);
-            return claim.getClaimStr();
-        } catch (SDKException e) {
-            throw new SDKException(ErrorCode.CreateOntIdClaimErr);
-        }
-    }
 
     public Map sortMap(Map<String, Object> claimMap) {
         Map<String, Object> contentMap = new HashMap();
@@ -880,46 +831,6 @@ public class OntIdTx {
         return mNew;
     }
 
-    /**
-     * verify OntId Claim
-     *
-     * @param claim
-     * @return
-     * @throws Exception
-     */
-    public boolean verifyOntIdClaim(String claim) throws Exception {
-        DataSignature sign = null;
-        try {
-
-            String[] obj = claim.split("\\.");
-            if (obj.length != 3) {
-                throw new SDKException(ErrorCode.ParamError);
-            }
-            byte[] payloadBytes = Base64.getDecoder().decode(obj[1].getBytes());
-            JSONObject payloadObj = JSON.parseObject(new String(payloadBytes));
-            String issuerDid = payloadObj.getString("iss");
-            String[] str = issuerDid.split(":");
-            if (str.length != 3) {
-                throw new SDKException(ErrorCode.DidError);
-            }
-            String issuerDdo = sdk.nativevm().ontId().sendGetDDO(issuerDid);
-            JSONArray owners = JSON.parseObject(issuerDdo).getJSONArray("Owners");
-            if (owners == null) {
-                throw new SDKException(ErrorCode.NotExistCliamIssuer);
-            }
-            byte[] signatureBytes = Base64.getDecoder().decode(obj[2]);
-            byte[] headerBytes = Base64.getDecoder().decode(obj[0].getBytes());
-            JSONObject header = JSON.parseObject(new String(headerBytes));
-            String kid = header.getString("kid");
-            String id = kid.split("#keys-")[1];
-            String pubkeyStr = owners.getJSONObject(Integer.parseInt(id) - 1).getString("Value");
-            sign = new DataSignature();
-            byte[] data = (obj[0] + "." + obj[1]).getBytes();
-            return sign.verifySignature(new Account(false, Helper.hexToBytes(pubkeyStr)), data, signatureBytes);
-        } catch (Exception e) {
-            throw new SDKException(ErrorCode.VerifyOntIdClaimErr);
-        }
-    }
 
     public Object getProof(String txhash) throws Exception {
         Map proof = new HashMap();
